requirements.txt
alabaster==0.7.10
asn1crypto==0.24.0
Babel==2.5.3
backoff==1.4.3
cassandra-driver==3.13.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
cryptography==2.1.4
docutils==0.14
idna==2.6
imagesize==0.7.1
inflection==0.3.1
Jinja2==2.10
MarkupSafe==1.0
mock==2.0.0
more-itertools==4.1.0
ndg-httpsclient==0.4.4
nose==1.3.7
numpy==1.14.0
packaging==16.8
pandas==0.22.0
pbr==3.1.1
pkg-resources==0.0.0
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
pymongo==3.6.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
Quandl==3.3.0
requests==2.18.4
retrying==1.3.3
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.0
sphinxcontrib-websupport==1.0.1
urllib3==1.22
